Update README.md
Fix spacing, steps in getting started numbered, add how it all works section, add version history, add how to contribute

How it all works (deprecated): An Arduino microcontroller is placed in the blaster. Through a series of inputs, the microcontroller can detect when a dart is fired. Depending on the mode, a switch, oriented behind the physical trigger of the blaster to be pressed with the trigger, will communicate to the microcontroller the compression of the blaster's trigger; the firing of a dart. In fully automatic blasters where this method may not be viable, an infrared (IR) receiver and transmitter pair will be integrated into the barrel of the blaster, with the transmitter continuously energizing the receiver by orienting the transmitter and receiver to be directly across through the diameter of the barrel. When a dart is fired, it will break this continuous energization of the receiver. This dramatic change in energization reading is detected by microcontroller.